Switch to MySQL and fix several issues

+++ app/ladder/models.py
@@ -166,22 +166,22 @@ class LadderSettings(SingletonModel):
 
 
 class DiscordChannels(SingletonModel):
-    polls = models.PositiveIntegerField(null=True, blank=True)
-    queues = models.PositiveIntegerField(null=True, blank=True)
-    chat = models.PositiveIntegerField(null=True, blank=True)
+    polls = models.BigIntegerField(null=True, blank=True)
+    queues = models.BigIntegerField(null=True, blank=True)
+    chat = models.BigIntegerField(null=True, blank=True)
 
 
 class DiscordPoll(models.Model):
     name = models.CharField(max_length=200)
-    message_id = models.PositiveIntegerField()
+    message_id = models.BigIntegerField()
 
 
 class QueueChannel(models.Model):
     name = models.CharField(max_length=200)
     min_mmr = models.PositiveSmallIntegerField(default=0)
     max_mmr = models.PositiveSmallIntegerField(default=0)
-    discord_id = models.PositiveIntegerField()
-    discord_msg = models.PositiveIntegerField(null=True, blank=True)
+    discord_id = models.BigIntegerField()
+    discord_msg = models.BigIntegerField(null=True, blank=True)
     record_matches = models.BooleanField(default=True)
 
     active = models.BooleanField(default=True)
@@ -217,7 +217,7 @@ class LadderQueue(models.Model):
 
     game_start_time = models.DateTimeField(null=True, blank=True)
     game_end_time = models.DateTimeField(null=True, blank=True)
-    game_server = models.PositiveIntegerField(null=True, blank=True)
+    game_server = models.BigIntegerField(null=True, blank=True)
 
     def __str__(self):
         return f'Queue #{self.id}'
